contacts_api/requirements.txt
fastapi
sqlalchemy
psycopg2-binary
uvicorn
pydantic
bcrypt
python-jose
passlib
